Game and GameLobby add ExitArea and detect player entering new GameLobby for detecting ready Player openUi use new type Ui.OpenUi OptionsManager remove obsolete TAA OptionsManager Add Master Bus Volume OptionsManager default AimCrosshairOpacity to 0.1 DifficultyConfig fix broken Serialize Menu add tween for opening and closing create OpenUi for async Free'ing, so exit animation tweens are possible MapSelect network selected map MapSelect show selected map on world view MapSelect add open/close animations OptionsMenu/MapSelectSelectMenu use CheckBox instead of CheckButton

+++ scripts/global/struct/DifficultyConfig.cs
@@ -61,10 +61,10 @@ public struct DifficultyConfig
     public readonly string Serialize()
     {
         var str = "";
-        str += Difficulty + ',';
-        str += Overscaling.ToString(".00") + ',';
-        str += (C1 ? '1' : '0') + ',';
-        str += (C2 ? '1' : '0') + ',';
+        str += (int)Difficulty + ",";
+        str += Overscaling.ToString(".00") + ",";
+        str += (C1 ? "1" : "0") + ",";
+        str += (C2 ? "1" : "0") + ",";
         str += MapDifficultyScale.ToString(".00");
         return str;
     }
